refactor: rename angular module to ngOvhContracts
