+++ CONTRIBUTING.md
@@ -1,4 +1,4 @@
-# Contributing to ovh-angular-contracts
+# Contributing to ng-ovh-contracts
 
 This project accepts contributions. In order to contribute, you should
 pay attention to a few things:
@@ -19,8 +19,8 @@ The contributions should be submitted through Github Pull Requests.
 
 # Licensing for new files
 
-ovh-angular-contracts is licensed under a BSD-3-Clause license. Anything
-contributed to ovh-angular-contracts must be released under this license.
+ng-ovh-contracts is licensed under a BSD-3-Clause license. Anything
+contributed to ng-ovh-contracts must be released under this license.
 
 When introducing a new file into the project, please make sure it has a
 copyright header making clear under which license it's being released.
