+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@ovh-ux/ovh-angular-contracts",
+  "name": "@ovh-ux/ng-ovh-contracts",
   "version": "3.0.0-beta.0",
   "description": "Component to correctly show contracts in OVH control panels",
   "keywords": [
@@ -7,14 +7,14 @@
     "contracts",
     "ovh"
   ],
-  "repository": "ovh-ux/ovh-angular-contracts",
+  "repository": "ovh-ux/ng-ovh-contracts",
   "license": "BSD-3-Clause",
   "author": "OVH SAS",
   "files": [
     "dist"
   ],
-  "main": "./dist/cjs/ovh-angular-contracts.js",
-  "browser": "./dist/umd/ovh-angular-contracts.js",
+  "main": "./dist/cjs/index.js",
+  "browser": "./dist/umd/ng-ovh-contracts.js",
   "scripts": {
     "build": "rollup -c",
     "dev": "rollup -c --watch",
